Android: Register AndroidDevice creator regardless of SDK state
The AndroidDeviceFactory only registered the AndroidDevice creator
function if the Android SDK was correctly set up. That check occurred
only once during the Qt Creator start up sequence.

As a consequence, Qt Creator had to be restarted after an initial set up
of the Android SDK. Users are not notified of such requirement, and we
actually don't want to enforce a restart.

With this change, the AndroidDevice creator function is always
registered, and the "Add Android Device" wizard always selectable. Users
who try to add an Android device without having set up Android before
get to see an info message box.

Change-Id: I0600e36575c2dd075af9398597c3c8dab3bb243d
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>