McuSupport: Read versioned settings keys for MCU dependencies
In an accompanying change, the installer writes versioned
settings keys. The version string will be appended to the
existing key, after a single underscore character.

When reading from the settings, the plugin first looks for
keys matching one of the requested versions from the kit.

Optionally, if no key for a matching version is found, the
key for the newest version available is chosen. This only
applies to the Qul SDK package for the time being.

If no suitable versioned key is found, the plain unversioned
settings key is picked

Task-number: QTCREATORBUG-29194
Change-Id: I2db888390cfb64a4b7c78ebcf795543251cb7a1b
Reviewed-by: Eike Ziller <eike.ziller@qt.io>