ScreenRecorder: Fix widget colors in the Crop&Trim dialog
panelWidget() and isInUnstyledDialogOrPopup() in ManhattanStyle.cpp
would mark the contents of this dialog as "panel widgets". In order to
evade this, we don't set the dialog type of the Crop&Trim dialog to
Qt::Windows.

The author of this dialog did not leave us a comment on why Qt::Windows
was chosen, so we can assume it was done for no particular reason.

Fixes: QTCREATORBUG-29893
Change-Id: I90f653e39cee839203b8767cb30eadab69baaf09
Reviewed-by: Cristian Adam <cristian.adam@qt.io>