ExternalTools: Fix drag and drop in the preferences
It mostly broke when Qt changed containers from int to qsizetype:
When we use QDataStream to serialize the value of `QList::indexOf`, but
deserialize that into an `int` variable, we don't get the same value
back.

Fix that, and also use begin/endMoveRows, which results in a better
selection behavior after dropping.

Fixes: QTCREATORBUG-30469
Change-Id: Ic99181ea7f75958766977ce7cf9d17c3d96103e2
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>