LinuxDeviceTester: Fix a crash on closing dialog when tryToConnect()
Make a call to tryToConnect() a yet another task inside the task tree.
Prolong the lifetime of the linux device until the tryToConnect()
running in a separate thread is finished. This fixes a crash
when canceling the test and closing the preferences dialog while
tryToConnect() is still running.

Amends e8ad29d3affd115573d7ee7a5621a6ca6e168911

Change-Id: I5aca6c9c3de0792421b9774faf45cd463854265e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: hjk <hjk@qt.io>