QmlDesigner: Add optional dependency on LicenseChecker
It is important that the LicenseChecker's initialization (if that is
available) happens before QmlDesigner's, so add an optional dependency
to reflect that.

Add a PLUGIN_MANUAL_DEPENDS option to add_qtc_plugin, which explicitly
specifies the name, version and type. The existing PLUGIN_RECOMMENDS
requires the plugin target to exist to extract the plugin's version.

Change-Id: Ie2cf84e75964ce91ed8bbcdbeee9fa9770bed641
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>