Use Qt::TextShowMnemonic for text metrics and drawing
Qt installations on KDE systems may insert shortcut ampersands into
widgets texts. For self-painted widgets, we need therefore consider
Qt::TextShowMnemonic when retrieving metrics of a text and when drawing
it.

This change addresses the issue for Utils::DetailsButton and
Core::Button. Most likely, there are more places where this needs to be
fixed.

Task-number: QTCREATORBUG-30534
Change-Id: I98ef9bf8c8502d92d3a3884d1588be190934f5bd
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>