Do not use original ID when creating kit for vanished target
If we use the same ID for the new kit as for the original, we get into
problems if the original kit re-appears.

This can happen when uninstalling a Qt version with the online
installer, and later re-installing it. Or when changing the install
settings path to a different online installation and back again via the
command line or the "Link with Qt" functionality.

Change-Id: Id99a3560ebed6264ad13dab5ff48945b30a8229f
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>