Spinner: Allow setting the spinner color
Deeply under the hood, the color for ther Spinner was determined by the
QApplication palette, color role "Text".

In order to run the spinner on a dark toolbar (in a non-dark theme),
though, the color needs to be bright. It is the case for the
ScreenRecorder.

This change adds a color setter to the Spinner classes, adds colored
Spinners to tst_spinner and makes use of the feature in the
ScreenRecorder.

The default color of the Spinner remains QApplication palette, color
role "Text".

Change-Id: Ia75758707f225b1f994e73c5a314478101c80f5e
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>