QmlDesigner: Fix landing page theming
The QmlDesigner landing page appears if a *.ui.qml document is open, and
user selects the "design" mode. Between QtC 12 and 13, the landing page
broke (unnoticed) due to the removal of the "Welcome_*" theme colors.

This change lets the Landing page use the new color tokens. A few fixes
in the PushButton component were needed.

Fixes: QTCREATORBUG-30547
Change-Id: I93319665bb283fba7f7a8892b671aaeb11479c7e
Reviewed-by: Eike Ziller <eike.ziller@qt.io>