Terminal: Ignore "height = 0" resize events
ConPTY on windows will send "clear line" characters in response
to resizing the PTY.

During re-showing the terminal pane the widget receives a "height=0"
resize event. We have to ignore this otherwise the conpty might try to
clear the screen.

Another issue was that we are calling resize on the pty even thought the
process has ended and won't be able to respond to the clear attempty by
the pty anymore.

Fixes: QTCREATORBUG-30523
Change-Id: I24caeaffb31d255a0640952e2d35bda23fd16280
Reviewed-by: Cristian Adam <cristian.adam@qt.io>