Doc: Auto-generate navigation links based on "All Topics"
This does not change what you see as values of \previouspage
and \nextpage in the source files. However, when you build the
docs, QDoc replaces those values with links to the next and
previous page from the list in qtdesignstudio-toc.qdoc because
that page contains "\title All Topics".

So, with this change, you need to keep the list in "All Topics"
up to date.

Change-Id: I0996e383cbd60b7f4297524e4141a2d7af6fae13
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Mats Honkamaa <mats.honkamaa@qt.io>