Fix visibility of vulnerability warning for QML debugging
The warning was not shown when opening a project with the QML debugging
option enabled.

The `changeHandler` was explicitly called at construction, but at this
point the `warningLabel` doesn't have a parent. The `setVisible` call
was skipped in this case, to prevent opening a toplevel window (for a
short time) for it until it has a parent.

Unfortunately there is no Qt API for setting the "visibility to the
parent". Instead invert the logic: Let the label have default visibility
to start with (which is "visible if the parent is visible"), and force
it to hidden in `changeHandler` if necessary.

Amends eda5fbd64575c0c268ba173e778a256a84484191

Change-Id: I69b86a967ee1fe7bd3d5d035765349981b64eb72
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>