Axivion: Improve path mapping
If the linked project is not the top level project our
path mappings did not work.
Try to get the correct file path by using QC internal
find functionality.
This helps for opening files by activating issues inside
the issues table or when clicking links of the issue
details.
Unfortunately this does not help for the inline
annotations or respective marks.

Change-Id: Ie34e1b20ff8b1b2b37e9f04c1d41bc2a4c33f260
Reviewed-by: Eike Ziller <eike.ziller@qt.io>