Utils: explicitly search the ClientExtras child object by name only
This class is missing the Q_OBJECT macro, so findChild was searching for
a QObject with the appropriate name, then casting to ClientExtras. That
means this change is a no-op.

But it makes explicit what we meant. Plus, findChild() and
findChildren() will enforce that the class they're searching for has the
Q_OBJECT macro.

Task-number: QTBUG-105023
Change-Id: I5f663c2f9f4149af84fefffd17c024995ef0c1af
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>