Editor: Open in different editor type after renaming or saving as
After save as, and when we get notified about a file rename, check if it
makes sense to open the document in a different editor type. This covers
the File > Save As menu item, as well as renaming in the project tree
and the file system view.

In principle this can happen when the MIME type changes (this could even
happen when the contents change, in special cases, but let's ignore
this).

Open the document in a different editor type if

- the document previously was opened in the default editor, and the new
file name would result in a different default editor type
- the document was not opened in the default editor, but the available
factories for the new name does not contain this

Do not use a different editor type, if it was previously open in a non-
default editor, and that is still valid for the new name.

Uses the openEditorWith functionality that is also used when a user
explicitly selects an editor type. This gives us all the details like
what happens if the new editor type fails to handle the file (e.g. when
renaming a text file to `.gif`).

Fixes: QTCREATORBUG-30317
Change-Id: I4c0b46da1e05165ac4f157dfa15b1783e0a3f3db
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>