Editor: move setScriptable to context actions
The scriptable info is saved to the action registered for a specific context
so we need to add this info to the action builder in the text editor that has the correct context.

Also remove all action builder constructs for the global context that have been registered before
in order to silence the add override action warning.

Change-Id: I631d71ba5eb8e975190713b33827926e868da295
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>