Set palette also for light/mixed (flat) themes if needed
We cannot assume that the default palette on the platform
and its theme fits the Qt Creator theme. If the system is dark, but Qt
Creator's theme is light, we need to set the theme palette on the
application.

Some themes already enforce the theme palette unconditionally (design-
light, and the dark themes). We don't want to set the palette
unconditionally for our flat default themes though, because for example
on macOS the light and dark colors that we get from the system are
always good and match the Qt Creator theme (because we tell the system
that our application is light/dark explicitly). Add a new
DerivePaletteFromThemeIfNeeded flag as an alternative to
DerivePaletteFromTheme for this.

Fixes: QTCREATORBUG-18281
Task-number: QTCREATORBUG-23187
Fixes: QTCREATORBUG-26817
Fixes: QTCREATORBUG-28589
Fixes: QTCREATORBUG-30138
Change-Id: Id0042980807fa2efd663b856a45255bb453227fd
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>