Update recent projects asynchronously
Avoid blocking Qt Creator whenever it gets active or projects change, if
some kind of network mapped path is in the list of recent projects. We
check for existence of recent projects, and that can take a long time in
that case, especially if the target is not connected. We already avoid
checking on explicit IDevices, but that doesn't help for other kinds of
mounted directories.

Maintain an "exists" state for each item and update that
asynchronously, informing all interested parties of changes with the
recentProjectsChanged signal. The state is saved in the settings,
meaning that projects that were not found previously start out hidden,
under the assumption that the existence check is fast for them if they
become available again.

Fixes: QTCREATORBUG-30681
Change-Id: Ic39a88b6b5128c3ae4582a6c66fc16be4b297e56
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>