AutoTest: Fix running order for QTest tests
Users can rely on the order of test function execution as
there is a fixed running order.
Fix the order by using location information we have instead
of the display order inside the tree view.
Fixes the order for running selected tests, failed tests or
tests inside the current file.

Fixes: QTCREATORBUG-30670
Change-Id: I5229424c3c8a2a7e030ef95c98d6fb602e54cc87
Reviewed-by: David Schulz <david.schulz@qt.io>