SquishTests: Make sure that button in Designer is renamed properly
This misused typeLines() to close the dialog by pressing "Return".
Although the entire name of the button should be selected at first,
I witnessed the new name being appended to the old one.

The new code uses replaceEditorContent() to explicit select the old
name and replace it with the new name. After that, it explicitly
presses "Return" to close the dialog.

Change-Id: I5ec946f57000396b70428e08ab795a1ee3eba91a
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>