Themes: Prevent Qt from changing the app palette behind our back
And remove the "GTK_THEME" hack

Even though Qt Creator now looked fine when started with a
light/mixed theme on a dark system theme, we still had to force
a light default palette on the application at startup.

The issue is that when the user switches the system theme
to dark during Qt Creator's runtime, Qt notices that and changes
the default application palette to dark.

Prevent that by setting a detached default palette on the
application on startup, even if we do not intend to apply custom
colors to it.

Task-number: QTCREATORBUG-28497
Change-Id: I7a43f108c2299967a14f382e52b62c31491e673d
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>