ExtensionSystem: Don't export test symbols without tests configured
We want the plugin manager to support the -test option etc even if
Qt Creator is not configured with WITH_TESTS, so 3rdparty plugins can
still be built and tested with a released Qt Creator.

But we do not want to export functions/classes that are only needed for
Qt Creator tests in that case.

When Qt Creator is configured with WITH_TESTS, the QtCreatorAPI adds the
WITH_TESTS define, and we used the same define for the first case above,
where we only want to add the -test option and corresponding API. Use a
different define EXTENSIONSYSTEM_WITH_TESTOPTION for this. The
EXTENSIONSYSTEM_TEST_EXPORT keeps checking the WITH_TESTS define.

Amends e5a4fdaa1b65beb0ca257ef698f685303b867802

Change-Id: I547f1fa7413e3bc452f71ae920c310a05e0193ae
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>