QmlDesigner: Update the Layouts doc
This patch updates the Layout document. Removing old tables
as they are now mostly in tooltips. Also, it tries to
structure the instructions and slightly fix the wording.
It adds an example for Qt Quick Layout workflow and explains
how properties should be used to adjust the components
in the Layout.

Fixes: QDS-12282
Change-Id: I10ec4d8f527c605632843b23007c54c87b22fd86
Reviewed-by: Mats Honkamaa <mats.honkamaa@qt.io>