Editors: Highlight the view that has focus
If there is more than one editor view, i.e. if the view is split or an
additional editor window open, paint a small line in the highlight color
below the editor tool bar to indicate which view is "current", and
therefore the target for opening files, and keystrokes (if it also has
focus).

Fixes: QTCREATORBUG-23654
Change-Id: I9950c133a6633c6e68943c038669895dce1dd7ef
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>