EditorManager: Keep track of current view
in sync with the current editor.

Before, we either tracked the current editor _or_ the current view if
there was no current editor. This made usage more complicated. Tracking
the current view in sync with a current editor makes setting these two
more complicated, but on the usage side are simpler. It also allows us
to just set the current view (which also sets the current editor to the
view's current editor), and to specifically signal changes of the current
view, and to keep a history of previous current views later.

Task-number: QTCREATORBUG-30408
Task-number: QTCREATORBUG-23654
Change-Id: I32967cea10972ca8b6939a84a0fbb396ca5c3721
Reviewed-by: David Schulz <david.schulz@qt.io>