FancyTabWidget: Remove an old, unnecessary hack
When clicking on a mode button, the emission of `currentChanged` was
delayed by an event. That was supposedly to "improve the responsiveness"
of the mode bar, but I don't see any difference, and _if_ changing the
mode widget was a performance issue, this wouldn't be the right place to
fix or work around it.

The hack creates issues, because it separates the `currentAboutToChange`
and `currentChanged` signals by an event processing.

Amends bd2ba2307b5695a3856c01ea478066e411ca6256

Change-Id: I53a7aa74fdcfb733a28574837ca9e95fc2e97f0b
Reviewed-by: David Schulz <david.schulz@qt.io>