EditorManager: Make sure the current view is visible
If an editor area is hidden, for example by unchecking the "Editor" view
in Debug mode, this might hide the current view. This is ugly, because
opening an editor in that state either opens the editor in an invisible
view, or in case of Debug mode with hidden "Editor" view it switches to
Edit mode. Which is detrimental if the user has a separate editor window
open that could be used instead.

When an editor area is hidden, try to set a current view that is visible
(which might be the same as before).

Fixes: QTCREATORBUG-30408
Change-Id: I0eb65fddbc857b6efb12972b861b9f44204bd715
Reviewed-by: David Schulz <david.schulz@qt.io>