QmlDesigner: Handle material type without using item library entries
Recent project structure changes invalidated material bundle metainfo
files, as we can't support both new and old structure with them.
They were only used for populating possible types in material editor,
so adjusted this functionality to only allow changing the type between
basic types. If the material is a custom component, then the only the
current type is displayed and the type combo box is disabled.

Fixes: QDS-12628
Change-Id: I07b1a482977f0ecc94c35f3d485302c85c6153f9
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>