ProjectExplorer: introduce workspace project
A workspace project can be utilized to open projects that either do not
have a project manager or which uses a project manager we do not support
natively. The workspace project basically scans the folder recursively
and adds all files and folders to the project tree.

The user can provide a project name in the .qtcreator/project.json
project file. It is also possible to exclude files and subdirectories
from the project via the project trees context menu.

Change-Id: If84a4d96c6217ae024af4e2869eb600e1faad372
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>