C++: Do not freeze Qt Creator while checking file sizes
Before actually indexing files, the C++ model checks the files against a
file size limit (if set, which is the default).

Do not iterate over all files and check their size in the main thread.
If the files are on a device, this operation is not fast. Move the
filtering to the parsing thread itself.

Change-Id: I2202cc44c28f38159ca593db2399dde30f95f9bd
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>