Editors: Update navigation history when closing with the tool button
The code path for the global actions (File menu, shortcut) were adding
the closed document to the history, but the editor view local action did
not. Move the call one level deeper, so it is done for both.

That fixes Window > Go Back under some circumstances in which it didn't open closed files again.

Change-Id: I74bdb02209b8ac7bb7714d9a827e3a0c30a4c9e9
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>