PE: Fix ProjectTreeWidget::expandedCount not to use 0 for error cases
This would make ProjectTreeWidget::nodeForFile pick the wrong node due
to nodeExpandCount having a lower count.

In case of qml files there is the source file node and the resource file
node pointing to the same file name, but the resource name has an invalid
index (not visible in the project view) and shouldn't be preffered!

Fixes: QTCREATORBUG-30571
Change-Id: I2aff7da9dd6d83222e341b5cbcb78db5cd6a9224
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>