LanguageClient: Let clients opt out of sending the didSave message
... and make use of it in ClangdClient.
When a file gets saved, clangd checks all open documents for whether they
are affected by that. While the answer is usually "no", there is still
one thread per document being started to find that out, which triggers
users who keep a close watch on CPU usage. To alleviate this, we now
send the didSave message only for header files.

Fixes: QTCREATORBUG-28580
Change-Id: Id761cec8b04e95a795c8bd8754466bf03b6f2c28
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: David Schulz <david.schulz@qt.io>