QmlDesigner: Add UniqueName class
The class is used for generating a unique name anywhere in QmlDesigner
codebase. The role of this class is generation of next candidate name
when name is not unique. Actual checking of name uniqueness is provided
as a function param.

Also applied the new UniqueName on 1 usecase in
ContentLibraryUserModel::getUniqueLib3DQmlName

Change-Id: I777aeef7c269bed7d999695cf5fcee6a5576222b
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>